# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/users": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "Create a new user",
        "operationId": "createUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/{userId}/rooms": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "Add a user to a room",
        "operationId": "addMember",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomId"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User added to room",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Failed to add user to room",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "Authenticate user and return JWT token",
        "operationId": "authenticate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Authenticated successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        }
      }
    },
    "/rooms": {
      "post": {
        "tags": [
          "room-controller"
        ],
        "summary": "Create a new room",
        "operationId": "createRoom",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Room created successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RoomResponse"
                }
              }
            }
          }
        }
      }
    },
    "/rooms/{roomId}": {
      "get": {
        "tags": [
          "room-controller"
        ],
        "summary": "Get a room by its ID",
        "operationId": "getRoomById",
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Room found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RoomResponse"
                }
              }
            }
          },
          "404": {
            "description": "Room not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RoomResponse"
                }
              }
            }
          }
        }
      }
    },
    "/rooms/{roomId}/users": {
      "get": {
        "tags": [
          "room-controller"
        ],
        "summary": "Get users in a room",
        "operationId": "getUsers",
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/rooms/{roomId}/messages": {
      "get": {
        "tags": [
          "room-controller"
        ],
        "summary": "Get messages for a room",
        "operationId": "getMessages",
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Messages retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/rooms/name/{name}": {
      "get": {
        "tags": [
          "room-controller"
        ],
        "summary": "Search rooms by name",
        "operationId": "getRoomByName",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Rooms retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoomResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/{userId}/rooms/{roomId}": {
      "delete": {
        "tags": [
          "user-controller"
        ],
        "summary": "Remove a user from a room",
        "operationId": "removeMember",
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User removed from room",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Failed to remove user from room",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UserRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "Email address as username",
            "example": "john.doe@example.com"
          },
          "password": {
            "pattern": "^(?=.*[0-9])(?=.*[A-Z]).{8,}$",
            "type": "string",
            "description": "Secure password",
            "example": "StrongPass123"
          }
        },
        "description": "User credentials used for login or registration"
      },
      "RoomResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "UUID of the room",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "description": "Name of the room",
            "example": "general"
          }
        },
        "description": "Response with basic room info"
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "Username (email)",
            "example": "john.doe@example.com"
          },
          "id": {
            "type": "integer",
            "description": "User ID",
            "format": "int64",
            "example": 42
          },
          "rooms": {
            "uniqueItems": true,
            "type": "array",
            "description": "Rooms the user has joined",
            "items": {
              "$ref": "#/components/schemas/RoomResponse"
            }
          }
        },
        "description": "User response containing basic user info and joined rooms"
      },
      "RoomId": {
        "type": "object",
        "properties": {
          "roomId": {
            "type": "string",
            "description": "UUID of the room",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        },
        "description": "Wrapper for room ID"
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "JWT token",
            "example": "eyJhbGciOiJIUzI1NiIsIn..."
          },
          "expiresIn": {
            "type": "integer",
            "description": "Token expiration time in seconds",
            "format": "int64",
            "example": 3600
          },
          "id": {
            "type": "integer",
            "description": "ID of the authenticated user",
            "format": "int64",
            "example": 42
          }
        },
        "description": "Response returned after successful login"
      },
      "RoomCreateRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the room",
            "example": "general"
          }
        },
        "description": "Request to create a new chat room"
      }
    }
  }
}